Upgrade all dependencies to fix type error
Couldn't deploy earlier due to some type errors, so I just upgraded
everything. This commit can be deployed (after a .env file is created).

The switch to Webpack 4 required a new webpack.config.js. This was
largely obtained from a serverless-webpack example, but adapted for
TypeScript.

Webpack config that we shamelessly adapted heavily:
https://github.com/serverless-heaven/serverless-webpack/blob/5cf24795fe1501e2e7c71dfe60368fc6578efd25/examples/babel-webpack-4/webpack.config.js